+++ rapidenv/osh/__init__.py
@@ -1,2 +1,3 @@
 from .osh import copy, run_process, run_process_with_stdout
 from .download import download_archive
+from .copy_large_files import copy_large